+++ src/main/java/kr/syeyoung/dungeonsguide/features/impl/etc/FeatureTooltipPrice.java
@@ -45,14 +45,14 @@ public class FeatureTooltipPrice extends SimpleFeature implements TooltipListene
                 public int compare(String o1, String o2) {
                     String id2 = id + "::" + o1 + "-" + enchants.getInteger(o1);
                     AhUtils.AuctionData auctionData = AhUtils.auctions.get(id2);
-                    int price1 = (auctionData == null) ? 0 : auctionData.lowestBin;
+                    long price1 = (auctionData == null) ? 0 : auctionData.lowestBin;
                     String id3 = id + "::" + o2 + "-" + enchants.getInteger(o2);
                     AhUtils.AuctionData auctionData2 = AhUtils.auctions.get(id3);
-                    int price2 = (auctionData2 == null) ? 0 : auctionData2.lowestBin;
+                    long price2 = (auctionData2 == null) ? 0 : auctionData2.lowestBin;
                     return (compare2(price1, price2) == 0) ? o1.compareTo(o2) : compare2(price1, price2);
                 }
 
-                public int compare2(int y, int x) {
+                public int compare2(long y, long x) {
                     return (x < y) ? -1 : ((x == y) ? 0 : 1);
                 }
             });
